requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argilla==1.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
bce-python-sdk==0.8.74
beautifulsoup4==4.11.2
black==21.4b2
bleach==6.0.0
blis==0.7.9
boto3==1.26.102
botocore==1.29.102
Brotli==1.0.9
cachetools==5.3.0
camelot-py==0.11.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
comm==0.1.2
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.7
cryptography==39.0.2
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
dataclasses-json==0.5.7
datasets==2.7.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
dill==0.3.6
dnspython==2.3.0
effdet==0.3.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.3.0
executing==1.2.0
fastai==2.7.12
fastapi==0.89.1
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.16.3
fastprogress==1.0.3
ffmpy==0.3.0
filelock==3.8.0
fire==0.5.0
Flask==2.2.2
flask-babel==3.0.1
flatbuffers==23.3.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
geojson==2.5.0
gevent==22.10.2
geventhttpclient==2.0.2
google-api-core==2.11.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-cloud-documentai==2.13.0
google-cloud-vision==3.4.0
googleapis-common-protos==1.58.0
gradio==3.17.1
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.25.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.0.126
langcodes==3.3.0
layoutparser==0.3.4
linkify-it-py==1.0.3
lmdb==1.4.0
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
onnxruntime==1.14.1
openai==0.27.2
opencv-contrib-python==4.7.0.68
opencv-python==4.6.0.66
openpyxl==3.1.0
opt-einsum==3.3.0
orjson==3.8.5
packaging==23.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.2
paddlepaddle==2.4.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pdf2docx==0.5.6
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
platformdirs==3.1.0
portalocker==2.7.0
premailer==3.10.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pydot==1.4.2
pydub==0.25.1
Pygments==2.14.0
PyMuPDF==1.20.2
pyowm==3.3.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.6.0
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.5
python-pptx==0.6.21
python-rapidjson==1.9
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
rapidfuzz==2.13.3
rarfile==4.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
seqeval==1.2.2
Shapely==1.8.5.post1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
srsly==2.4.6
stack-data==0.6.2
starlette==0.22.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
terminado==0.17.1
thinc==8.1.9
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.11.0
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@19ade2426a6ec12cfbec2779f32573e7b02d49a0
tritonclient==2.30.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
unstructured==0.5.7
unstructured-inference==0.2.14
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.20.0
visualdl==2.5.0
Wand==0.6.11
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.14.1
x2paddle==1.4.0
XlsxWriter==3.0.9
xxhash==3.1.0
yacs==0.1.8
yarl==1.8.1
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
